Replaces the resources form and provides are btn for create new
Needed to modify the logic in the resources form and had basically
rewritten the whole snippet. No more ckan_extends, this replaces the
whole file.

In this (intermediate commit) I have included a button for reloading the
page with a complete new resource for editing if that's how the form is
retrieved. If the form is not retrieved for new_resource then the button
does not appear. I want The button to be hidden and trigger
automatically, so that's the next step.