Adds JSON info for ddict and forces url to be dummy when absent
If url doesn't exist then changing resources will prompt an error and
some resources don't show the url so it can't be edited. Now when
resource changes the error stil appears, but the resource populates the
URL so after 2 updates the new resource_type can be committed.
Definitely not ideal, and a better fix might be changing url validation
to forcibly create a dummy url